test: fix deterministic parser seed type

+++ app/src/test/java/com/charmingcolor/shuttersoundzero/core/adb/AdbShellCommandResultParserTest.kt
@@ -142,7 +142,7 @@ class AdbShellCommandResultParserTest {
 
     @Test
     fun parseOrNull_forRandomMarkerLikeNoise_neverTreatsPartialLinesAsCompletion() {
-        val random = Random(0xF00DCAFE)
+        val random = Random(0xF00DCAFE.toInt())
         val separators = listOf(" ", "\t", "-", "+", ":", "x")
 
         repeat(512) { index ->
